% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Anfis-getters.R
\docType{methods}
\name{getRules}
\alias{getConsequents}
\alias{getConsequents,ANFIS-method}
\alias{getErrors}
\alias{getErrors,ANFIS-method}
\alias{getPremises}
\alias{getPremises,ANFIS-method}
\alias{getPremises-methods}
\alias{getRules}
\alias{getRules,ANFIS-method}
\alias{getRules-methods}
\alias{getTrainingType}
\alias{getTrainingType,ANFIS-method}
\title{Getters for ANFIS object}
\usage{
getRules(object)

\S4method{getRules}{ANFIS}(object)

getPremises(object)

\S4method{getPremises}{ANFIS}(object)

getConsequents(object)

\S4method{getConsequents}{ANFIS}(object)

getErrors(object)

\S4method{getErrors}{ANFIS}(object)

getTrainingType(object)

\S4method{getTrainingType}{ANFIS}(object)
}
\arguments{
\item{object}{ANFIS class object}
}
\value{
according to the call one of the following objects can be returned
 \item{matrix}{numeric matrix with rules or consequents}
 \item{list}{list with MembershipFunctions or premises and consequents}
 \item{character}{name of the trainingType}
 \item{numeric}{numeric vector with trainnig errors, fitted training values
   and residuals}
}
\description{
Obtain ANFIS's slot information, according to the given function call.


















}
\note{
see full example in \code{\link{ANFIS-class}}
}
\author{
Cristobal Fresno \email{cfresno@bdmg.com.ar}, Andrea S. Llera
 \email{ALlera@leloir.org.ar} and Elmer A. Fernandez
 \email{efernandez@bdmg.com.ar}
}
\seealso{
Other ANFIS: \code{\link{ANFIS-class}}; \code{\link{LSE}},
  \code{\link{LSE}}, \code{\link{LSE,ANFIS-method}},
  \code{\link{LSE-methods}},
  \code{\link{trainHybridJangOffLine}},
  \code{\link{trainHybridJangOffLine}},
  \code{\link{trainHybridJangOffLine,ANFIS-method}},
  \code{\link{trainHybridJangOffLine-methods}},
  \code{\link{trainHybridJangOnLine}},
  \code{\link{trainHybridJangOnLine}},
  \code{\link{trainHybridJangOnLine,ANFIS-method}},
  \code{\link{trainHybridJangOnLine-methods}},
  \code{\link{trainHybridOffLine}},
  \code{\link{trainHybridOffLine}},
  \code{\link{trainHybridOffLine,ANFIS-method}},
  \code{\link{trainHybridOffLine-methods}};
  \code{\link{anfis3}}; \code{\link{coef}},
  \code{\link{coef,ANFIS-method}},
  \code{\link{coefficients}},
  \code{\link{coefficients,ANFIS-method}},
  \code{\link{fitted}}, \code{\link{fitted,ANFIS-method}},
  \code{\link{fitted.values}},
  \code{\link{fitted.values,ANFIS-method}},
  \code{\link{resid}}, \code{\link{resid,ANFIS-method}},
  \code{\link{residuals}},
  \code{\link{residuals,ANFIS-method}},
  \code{\link{summary}},
  \code{\link{summary,ANFIS-method}};
  \code{\link{initialize}},
  \code{\link{initialize,ANFIS-method}};
  \code{\link{plotMF}}, \code{\link{plotMF}},
  \code{\link{plotMF,ANFIS-method}},
  \code{\link{plotMF-methods}}, \code{\link{plotMFs}},
  \code{\link{plotMFs}},
  \code{\link{plotMFs,ANFIS-method}},
  \code{\link{plotMFs-methods}}; \code{\link{plot}},
  \code{\link{plot,ANFIS-method}}; \code{\link{predict}},
  \code{\link{predict,ANFIS-method}}; \code{\link{print}},
  \code{\link{print,ANFIS-method}}, \code{\link{show}},
  \code{\link{show,ANFIS-method}}; \code{\link{trainSet}}
}

