% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Anfis-metrics.R
\docType{methods}
\name{fitted}
\alias{coef}
\alias{coef,ANFIS-method}
\alias{coefficients}
\alias{coefficients,ANFIS-method}
\alias{fitted}
\alias{fitted,ANFIS-method}
\alias{fitted.values}
\alias{fitted.values,ANFIS-method}
\alias{resid}
\alias{resid,ANFIS-method}
\alias{residuals}
\alias{residuals,ANFIS-method}
\alias{summary}
\alias{summary,ANFIS-method}
\title{ANFIS training results}
\usage{
\S4method{fitted}{ANFIS}(object, ...)

\S4method{fitted.values}{ANFIS}(object, ...)

\S4method{coef}{ANFIS}(object, ...)

\S4method{coefficients}{ANFIS}(object, ...)

\S4method{resid}{ANFIS}(object, ...)

\S4method{residuals}{ANFIS}(object, ...)

\S4method{summary}{ANFIS}(object, ...)
}
\arguments{
\item{object}{ANFIS class object}

\item{...}{required by resid, residuals, coef and coefficients}
}
\value{
according to the call one of the following objects can be returned
 \item{list}{list with premises and consequents.}
 \item{numeric}{numeric vector with training errors, fitted training values
   and residuals.}
 \item{printed}{statistics of the training process.}
}
\description{
Obtain ANFIS slot information, according to training output












}
\note{
see full example in \code{\link{ANFIS-class}}
}
\author{
Cristobal Fresno \email{cfresno@bdmg.com.ar}, Andrea S. Llera
 \email{ALlera@leloir.org.ar} and Elmer A. Fernandez
 \email{efernandez@bdmg.com.ar}
}
\seealso{
Other ANFIS: \code{\link{ANFIS-class}}; \code{\link{LSE}},
  \code{\link{LSE}}, \code{\link{LSE,ANFIS-method}},
  \code{\link{LSE-methods}},
  \code{\link{trainHybridJangOffLine}},
  \code{\link{trainHybridJangOffLine}},
  \code{\link{trainHybridJangOffLine,ANFIS-method}},
  \code{\link{trainHybridJangOffLine-methods}},
  \code{\link{trainHybridJangOnLine}},
  \code{\link{trainHybridJangOnLine}},
  \code{\link{trainHybridJangOnLine,ANFIS-method}},
  \code{\link{trainHybridJangOnLine-methods}},
  \code{\link{trainHybridOffLine}},
  \code{\link{trainHybridOffLine}},
  \code{\link{trainHybridOffLine,ANFIS-method}},
  \code{\link{trainHybridOffLine-methods}};
  \code{\link{anfis3}}; \code{\link{getConsequents}},
  \code{\link{getConsequents}},
  \code{\link{getConsequents,ANFIS-method}},
  \code{\link{getConsequents,ANFIS-method}},
  \code{\link{getErrors}}, \code{\link{getErrors}},
  \code{\link{getErrors,ANFIS-method}},
  \code{\link{getErrors,ANFIS-method}},
  \code{\link{getPremises}}, \code{\link{getPremises}},
  \code{\link{getPremises,ANFIS-method}},
  \code{\link{getPremises-methods}},
  \code{\link{getRules}}, \code{\link{getRules}},
  \code{\link{getRules,ANFIS-method}},
  \code{\link{getRules-methods}},
  \code{\link{getTrainingType}},
  \code{\link{getTrainingType}},
  \code{\link{getTrainingType,ANFIS-method}},
  \code{\link{getTrainingType,ANFIS-method}};
  \code{\link{initialize}},
  \code{\link{initialize,ANFIS-method}};
  \code{\link{plotMF}}, \code{\link{plotMF}},
  \code{\link{plotMF,ANFIS-method}},
  \code{\link{plotMF-methods}}, \code{\link{plotMFs}},
  \code{\link{plotMFs}},
  \code{\link{plotMFs,ANFIS-method}},
  \code{\link{plotMFs-methods}}; \code{\link{plot}},
  \code{\link{plot,ANFIS-method}}; \code{\link{predict}},
  \code{\link{predict,ANFIS-method}}; \code{\link{print}},
  \code{\link{print,ANFIS-method}}, \code{\link{show}},
  \code{\link{show,ANFIS-method}}; \code{\link{trainSet}}
}

