% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Anfis-plot.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,ANFIS-method}
\title{Plot ANFIS training errors}
\usage{
\S4method{plot}{ANFIS}(x, y, epoch = Inf, ...)
}
\arguments{
\item{x}{ANFIS class object.}

\item{y}{not used but necessary for redefining the generic function.}

\item{epoch}{for on-line only: epoch == Inf the whole training error;
epoch == integer > 0 the give epoch trainings errors, epoch == 0 the abs
epoch training sum of errors.}

\item{...}{plot additional parameters.}
}
\value{
output graphics.
}
\description{
Plot the training error of the network. If trainingType is "on-line" then
full pattern errors along the patterns of the whole training process; for
a specific epoch or the epoch summary error.
}
\note{
see full example in \code{\link{ANFIS-class}}
}
\author{
Cristobal Fresno \email{cfresno@bdmg.com.ar}, Andrea S. Llera
 \email{ALlera@leloir.org.ar} and Elmer A. Fernandez
 \email{efernandez@bdmg.com.ar}
}
\seealso{
\code{\link{ANFIS-class}}

Other ANFIS: \code{\link{ANFIS-class}}; \code{\link{LSE}},
  \code{\link{LSE}}, \code{\link{LSE,ANFIS-method}},
  \code{\link{LSE-methods}},
  \code{\link{trainHybridJangOffLine}},
  \code{\link{trainHybridJangOffLine}},
  \code{\link{trainHybridJangOffLine,ANFIS-method}},
  \code{\link{trainHybridJangOffLine-methods}},
  \code{\link{trainHybridJangOnLine}},
  \code{\link{trainHybridJangOnLine}},
  \code{\link{trainHybridJangOnLine,ANFIS-method}},
  \code{\link{trainHybridJangOnLine-methods}},
  \code{\link{trainHybridOffLine}},
  \code{\link{trainHybridOffLine}},
  \code{\link{trainHybridOffLine,ANFIS-method}},
  \code{\link{trainHybridOffLine-methods}};
  \code{\link{anfis3}}; \code{\link{coef}},
  \code{\link{coef,ANFIS-method}},
  \code{\link{coefficients}},
  \code{\link{coefficients,ANFIS-method}},
  \code{\link{fitted}}, \code{\link{fitted,ANFIS-method}},
  \code{\link{fitted.values}},
  \code{\link{fitted.values,ANFIS-method}},
  \code{\link{resid}}, \code{\link{resid,ANFIS-method}},
  \code{\link{residuals}},
  \code{\link{residuals,ANFIS-method}},
  \code{\link{summary}},
  \code{\link{summary,ANFIS-method}};
  \code{\link{getConsequents}},
  \code{\link{getConsequents}},
  \code{\link{getConsequents,ANFIS-method}},
  \code{\link{getConsequents,ANFIS-method}},
  \code{\link{getErrors}}, \code{\link{getErrors}},
  \code{\link{getErrors,ANFIS-method}},
  \code{\link{getErrors,ANFIS-method}},
  \code{\link{getPremises}}, \code{\link{getPremises}},
  \code{\link{getPremises,ANFIS-method}},
  \code{\link{getPremises-methods}},
  \code{\link{getRules}}, \code{\link{getRules}},
  \code{\link{getRules,ANFIS-method}},
  \code{\link{getRules-methods}},
  \code{\link{getTrainingType}},
  \code{\link{getTrainingType}},
  \code{\link{getTrainingType,ANFIS-method}},
  \code{\link{getTrainingType,ANFIS-method}};
  \code{\link{initialize}},
  \code{\link{initialize,ANFIS-method}};
  \code{\link{plotMF}}, \code{\link{plotMF}},
  \code{\link{plotMF,ANFIS-method}},
  \code{\link{plotMF-methods}}, \code{\link{plotMFs}},
  \code{\link{plotMFs}},
  \code{\link{plotMFs,ANFIS-method}},
  \code{\link{plotMFs-methods}}; \code{\link{predict}},
  \code{\link{predict,ANFIS-method}}; \code{\link{print}},
  \code{\link{print,ANFIS-method}}, \code{\link{show}},
  \code{\link{show,ANFIS-method}}; \code{\link{trainSet}}
}

