% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MembershipFunction-evaluateMF.R
\docType{methods}
\name{evaluateMF}
\alias{evaluateMF}
\alias{evaluateMF,BellMF-method}
\alias{evaluateMF,GaussianMF-method}
\alias{evaluateMF,MembershipFunction-method}
\alias{evaluateMF,NormalizedGaussianMF-method}
\alias{evaluateMF-methods}
\title{\code{evaluateMF} evaluate membership}
\usage{
evaluateMF(object, x)

\S4method{evaluateMF}{MembershipFunction}(object, x)

\S4method{evaluateMF}{BellMF}(object, x)

\S4method{evaluateMF}{GaussianMF}(object, x)

\S4method{evaluateMF}{NormalizedGaussianMF}(object, x)
}
\arguments{
\item{object}{MembershipFunction class heirs}

\item{x}{numeric of the MembershipFunction to be evaluated}
}
\value{
0 <= numeric <=1 with the obtained membership value
}
\description{
Evaluate de membership of x to the object MembershipFunction heirs.








}
\examples{
#BellMF example I
#A bell membership function with default prototype (a=1, b=1,c=0)
#The membership of x in the bell, should be 1
#The derivate of the first parameter at x, should be 0
#The derivate of the first parameter at x, should be also 0
bell <- new(Class="BellMF")
bell
evaluateMF(object=bell, x=0)
derivateMF(object=bell, x=0, i=1)
derivateMF(object=bell, x=0, i="a")
#
#BellMF example II
#A bell membership function with parameters (a=4,b=1,c=-10)
#The membership of x in the bell, should be 0.137931
#The derivate of the first parameter at x, should be 0.05945303
#The derivate on "a" at x=0, should be 0.05945303
bell2 <- new(Class="BellMF",parameters=c(a=4,b=1,c=-10))
bell2
evaluateMF(object=bell2, x=0)
derivateMF(object=bell2, x=0, i=1)
derivateMF(object=bell2, x=0, i="a")
#GaussianMF example I
#A Gaussian membership function with default prototype (mu=0, sigma=1)
#The membership of x in the Gaussian, should be 1/sqrt(2*pi) = 0.3989423
#The derivate of the first parameter at x, should be 0
#The derivate on "mu" parameter at x, should be 0
gaussian <- new(Class="GaussianMF")
gaussian
evaluateMF(object=gaussian, x=0)
derivateMF(object=gaussian, x=0, i=1)
derivateMF(object=gaussian, x=0, i="mu")
#
#GaussianMF example II
#A Gaussian membership function with paramateres (mu=0, sigma=1)
#The membership of x in the gaussian, should be 1/sqrt(2*pi) = 0.3989423
#The derivate of the first parameter at x, should be 0
#The derivate on "mu" parameter at x, should be 0
gaussian2 <- new(Class="GaussianMF",parameters=c(mu=0,sigma=1))
gaussian2
evaluateMF(object=gaussian2, x=0)
derivateMF(object=gaussian2, x=0, i=1)
derivateMF(object=gaussian2, x=0, i="mu")
#NormalizedGaussianMF example I
#A normalized Gaussian membership function with default paramateres (mu=0, sigma=1)
#The derivate of the first parameter at x, should be 1
#The derivate of the first parameter at x, should be 0
#The derivate on "mu" parameter at x, should be 0
normalizedGaussian <- new(Class="NormalizedGaussianMF")
normalizedGaussian
evaluateMF(object=normalizedGaussian, x=0)
derivateMF(object=normalizedGaussian, x=0, i=1)
derivateMF(object=normalizedGaussian, x=0, i="mu")
#
#NormalizedGaussianMF example II
#A normalized Gaussian membership function with paramateres (mu=0, sigma=1)
#' #The derivate of the first parameter at x, should be 1
#The derivate of the first parameter at x, should be 0
#The derivate on "mu" parameter at x, should be 0
normalizedGaussian2 <- new(Class="NormalizedGaussianMF",
 parameters=c(mu=0,sigma=1))
normalizedGaussian2
evaluateMF(object=normalizedGaussian2, x=0)
derivateMF(object=normalizedGaussian2, x=0, i=1)
derivateMF(object=normalizedGaussian2, x=0, i="mu")
}
\author{
Cristobal Fresno \email{cfresno@bdmg.com.ar}, Andrea S. Llera
 \email{ALlera@leloir.org.ar} and Elmer A. Fernandez
 \email{efernandez@bdmg.com.ar}
}
\seealso{
\code{\link{MembershipFunction-class}} and \code{\link{derivateMF}}

Other Membership Functions: \code{\link{BellMF}},
  \code{\link{BellMF-class}}; \code{\link{GaussianMF}},
  \code{\link{GaussianMF-class}};
  \code{\link{MembershipFunction}},
  \code{\link{MembershipFunction-class}};
  \code{\link{NormalizedGaussianMF}},
  \code{\link{NormalizedGaussianMF-class}};
  \code{\link{[,MembershipFunction-method}},
  \code{\link{[<-,MembershipFunction-method}},
  \code{\link{extract-methods}},
  \code{\link{extract-methods}}; \code{\link{derivateMF}},
  \code{\link{derivateMF}}, \code{\link{derivateMF}},
  \code{\link{derivateMF}}, \code{\link{derivateMF}},
  \code{\link{derivateMF,BellMF-method}},
  \code{\link{derivateMF,GaussianMF-method}},
  \code{\link{derivateMF,MembershipFunction-method}},
  \code{\link{derivateMF,NormalizedGaussianMF-method}},
  \code{\link{derivateMF-methods}};
  \code{\link{print,MembershipFunction-method}};
  \code{\link{show,MembershipFunction-method}}
}

