% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MembershipFunction-getset.R
\docType{methods}
\name{extract-methods}
\alias{[,MembershipFunction-method}
\alias{[<-,MembershipFunction-method}
\alias{extract-methods}
\title{Modify membership function parameters}
\usage{
\S4method{[}{MembershipFunction}(x, i)

\S4method{[}{MembershipFunction}(x, i) <- value
}
\arguments{
\item{x}{MembershipFunction class heirs}

\item{i}{numeric or character to access parameters vector [i]}

\item{value}{numeric parameter/s values}
}
\value{
\item{numeric}{parameter/s in the case of object[i]}
 \item{object}{MembershipFunction object in the case of object[i]<- value}
}
\description{
Get/set membership function parameters.


}
\seealso{
\code{\link{MembershipFunction-class}}

Other Membership Functions: \code{\link{BellMF}},
  \code{\link{BellMF-class}}; \code{\link{GaussianMF}},
  \code{\link{GaussianMF-class}};
  \code{\link{MembershipFunction}},
  \code{\link{MembershipFunction-class}};
  \code{\link{NormalizedGaussianMF}},
  \code{\link{NormalizedGaussianMF-class}};
  \code{\link{derivateMF}}, \code{\link{derivateMF}},
  \code{\link{derivateMF}}, \code{\link{derivateMF}},
  \code{\link{derivateMF}},
  \code{\link{derivateMF,BellMF-method}},
  \code{\link{derivateMF,GaussianMF-method}},
  \code{\link{derivateMF,MembershipFunction-method}},
  \code{\link{derivateMF,NormalizedGaussianMF-method}},
  \code{\link{derivateMF-methods}};
  \code{\link{evaluateMF}}, \code{\link{evaluateMF}},
  \code{\link{evaluateMF}}, \code{\link{evaluateMF}},
  \code{\link{evaluateMF}},
  \code{\link{evaluateMF,BellMF-method}},
  \code{\link{evaluateMF,GaussianMF-method}},
  \code{\link{evaluateMF,MembershipFunction-method}},
  \code{\link{evaluateMF,NormalizedGaussianMF-method}},
  \code{\link{evaluateMF-methods}};
  \code{\link{print,MembershipFunction-method}};
  \code{\link{show,MembershipFunction-method}}
}

